mnt: Clean roots yard from criu process
So here it is. If root task dies on restore the roots yard
dir remains unrmdired :( Since we already know its name, we
can remove one from criu. By the time we get to this place
the sub mount namespace(s) are already dead and yard dir
is empty. But umounting should be done by tasks after
successfull restore, so keep depopulation there.

+++ mount.c
@@ -2734,12 +2734,17 @@ int depopulate_roots_yard(void)
 		pr_perror("Can't unmount %s", mnt_roots);
 		ret = 1;
 	}
-	if (rmdir(mnt_roots)) {
-		pr_perror("Can't remove the directory %s", mnt_roots);
-		ret = 1;
-	}
 
 	return ret;
+}
+
+void cleanup_mnt_ns(void)
+{
+	if (mnt_roots == NULL)
+		return;
+
+	if (rmdir(mnt_roots))
+		pr_perror("Can't remove the directory %s", mnt_roots);
 }
 
 int prepare_mnt_ns(void)
